+++ mods/zora-create/src/loading.ts
@@ -1,0 +1,13 @@
+import { ModElement } from "@mod-protocol/core";
+
+const loading: ModElement[] = [
+  {
+    type: "text",
+    label: "Uploading the file: {{refs.myZoraCreateRequest.progress}}%",
+  },
+  {
+    type: "circular-progress",
+  },
+];
+
+export default loading;