+++ packages/react/src/index.tsx
@@ -176,10 +176,6 @@ const WrappedHorizontalLayoutRenderer = <T extends React.ReactNode>(props: {
 }) => {
   const { component: Component, element } = props;
   const { type, elements, ...rest } = element;
-
-  // Prevents the onLoad event from being called multiple times when
-  // the tree changes and reference to the events object changes.
-  // Assumes events are immutable over the lifecycle of a component
   const [events] = React.useState(element.events);
 
   React.useEffect(() => {
@@ -195,8 +191,6 @@ const WrappedVerticalLayoutRenderer = <T extends React.ReactNode>(props: {
 }) => {
   const { component: Component, element } = props;
   const { type, elements, ...rest } = element;
-
-  // See WrappedHorizontalLayoutRenderer for explanation
   const [events] = React.useState(element.events);
 
   React.useEffect(() => {